pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soundpulse"
version = "0.1.0"
description = "Music intelligence API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "python-Levenshtein>=0.26.0",
    "numpy>=2.1.0",
    "httpx>=0.28.0",
    "psycopg2-binary>=2.9.0",
    "rapidfuzz>=3.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "celery>=5.4",
    "pyjwt>=2.8",
    "cryptography>=42.0",
    "beautifulsoup4>=4.12",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
include = ["api*", "shared*", "scripts*", "scrapers*", "ml*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
